externalize WiFi creds and server host
