+++ ESP32/CarProtect_ESP32/config.h.example
@@ -1,0 +1,9 @@
+#pragma once
+
+// Rename this file to config.h and fill in your WiFi credentials
+// and the host name or IP address of the web server.
+
+#define WIFI_SSID "YOUR_WIFI_SSID"
+#define WIFI_PASSWORD "YOUR_WIFI_PASSWORD"
+#define SERVER_HOST "your.server.com"
+
