+++ README.md
@@ -1,22 +1,65 @@
-*Team Members*
+# CarProtect
 
-Girish Krishnan, Anish Sharma, Ian Tanuwidjaja, and Hamza Khan
-___
+CarProtect is a proof-of-concept IoT security system built during ECE140B.
+It combines two ESP32 based microcontrollers with a small web application
+that stores sensor data and displays live images.
 
-# ECE 140B: CarProtect
+## Repository layout
 
-Hello! This repository contains all the code files for our project. The complete demo video can be found at: https://youtu.be/MSkuZj5Wf60 
+```
+ESP32/              Firmware for the ESP32 boards
+  CamUpload/        Sketch for the ESP32-CAM module
+  CameraWebServerGirish/ Example camera project (reference)
+  CarProtect_ESP32/ Sketch for the main ESP32 board
 
-___
+web/                Python web application and Docker config
+  src/app/          Application package (routes, DB helpers)
+  src/public/       Static files served by the web server
+  src/templates/    Jinja2 templates
+```
 
-## How the Code is Organized + Instructions to Run the Code
+## Running the web server
 
-The __ESP 32__ folder contains the Arduino code that needs to be uploaded to the board. The *CameraWebServerGirish* code is not used in this project per se, but is a useful tutorial on using the ESP32 cam, that inspired us to build the rest of the project.
+The easiest way to run the server and database is via Docker:
 
-Inside the __ESP 32__ folder, you can find the __CamUpload__ folder, which contains the code that needs to be uploaded to the ESP32 Cam. The __CarProtect_ESP32__ folder contains the code that needs to be uploaded to the ESP32 Wrover Module board.
+```bash
+docker-compose up --build
+```
 
-Once the code is uploaded to each board, remove all connections to the ESP32 cam. Then, connect the 5V pin of the ESP32 cam to the 5V-OUT pin of the ESP32 Wrover Module, and connect the ground pin of the ESP32 cam to the ground pin of the Wrover module. Then, connect the ESP32 Wrover module to a power source.
+This starts a MySQL container and the Pyramid web application.
+The application listens on port `80` on the host and connects to
+MySQL using the credentials found in `credentials.env`.
 
-The __web__ folder contains the code necessary for the web server and database used for our full-stack IoT project. This involves the basic server architecture we learned in ECE 140A, and we have a Docker container with the web server + MySQL database. This is hosted via DigitalOcean.
+To run the application locally without Docker, install dependencies
+from `web/src/requirements.txt` and execute:
 
-Link: http://164.92.89.132/
+```bash
+python -m app.server
+```
+
+(You must set the environment variables defined in `credentials.env`.)
+
+`app.server` also respects two optional environment variables:
+
+* `APP_HOST` - interface to bind (default `0.0.0.0`)
+* `APP_PORT` - port number (default `6000`)
+
+## Firmware
+
+The `ESP32/CamUpload` sketch uploads images from the ESP32-CAM to the
+server. `ESP32/CarProtect_ESP32` reads sensor data (PIR, vibration and
+accelerometer) and periodically sends it to the server.
+Flash each sketch to the appropriate board using the Arduino IDE or
+`arduino-cli`.
+
+Before building, copy `config.h.example` from each sketch directory to
+`config.h` and fill in your WiFi credentials and the hostname or IP
+address of the web server.
+
+After flashing, connect the ESP32-CAM's `5V` and `GND` pins to the
+corresponding pins on the main ESP32 board and power the system.
+
+## Demo
+
+A short demonstration video can be found at
+<https://youtu.be/MSkuZj5Wf60>.
